Fixed the iFort crash for div runs
iFort apparently cannot properly allocate a composite type with
internal allocated parts when this type is declared as a private
variable in an OpenMP loop.  I don't know if it is supposed to -
gfortran handles this situation just fine.

Anyway, I moved the allocation into the loop.  This fixes the problem
with iFort+OpenMP. It is a bit wasteful for a linear code, but that
is worth considering the speed gains of parallel code