Change the interpretation of the second MMF argument
The second MMF argument is now DFRAC-OR-FILL.  When larger than 1, it
is interpreted as the fractal dimension.  When less than 1, it is
interpreted as the filling factor for the aggregate.

The advantage is that we can keep the porosity as something that
affect the monomers and does not have a double meaning depending on
the details of the MMF arguments -  that was pretty confusing.