Implement the possibility to force nonzero matrix elements
When MFF detects that the phase shift is too large, we do not have a
reliable way to compute the scattering matrix for short wavelengths.
The user can now use -mmfss instead -mmf as the method.  Then, at
short wavelengths where the phase shift is too large, we return the
scattering matrix computed under the assumption of single scattering
in the aggregates, just for those wavelengths where the phase shift is
too large.  This may be reasonable for absorming materials, but we do
not have a good criterion to say for sure.  There are plenty of
warnings for the user hen this option is used.